fixed nextcloud ffmpeg install script to use dpkg instead of pacman (pacman is not present in nextcloud container), and added vi to pacstrap command in arch_install.sh

+++ system/server/nextcloud/nextcloud-install-ffmpeg.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
-docker exec nextcloud apt -y update
-docker exec nextcloud apt -y install libmagickcore-6.q16-6-extra ffmpeg
+if ! dpkg -s ffmpeg &> /dev/null; then
+  # Install ffmpeg if not present
+  docker exec nextcloud apt -y update
+  docker exec nextcloud apt -y install libmagickcore-6.q16-6-extra ffmpeg
+fi
